refactor: Moved packages and add package-info.java
