+++ src/main/java/fr/siamois/domain/package-info.java
@@ -1,0 +1,7 @@
+/**
+ * This package contains code that is related to the domain of the application.
+ * Any code within this package is business code and only concerns the application.
+ *
+ * @author Julien Linget
+ */
+package fr.siamois.domain;
