+++ src/main/java/fr/siamois/infrastructure/package-info.java
@@ -1,0 +1,7 @@
+/**
+ * This package contains code that is related to the infrastructure of the application.
+ * Any code within the package contains persistant data related code. (Database, Api, Files, ...)
+ *
+ * @author Julien Linget
+ */
+package fr.siamois.infrastructure;
