+++ src/main/java/fr/siamois/ui/config/api/OpenApiConfig.java
@@ -1,4 +1,4 @@
-package fr.siamois.domain.config.api;
+package fr.siamois.ui.config.api;
 
 import io.swagger.v3.oas.annotations.OpenAPIDefinition;
 import io.swagger.v3.oas.annotations.enums.SecuritySchemeIn;
