+++ src/main/java/fr/siamois/ui/package-info.java
@@ -1,0 +1,7 @@
+/**
+ * This package contains code that is related to the direct user interactions with the applications.
+ * Any code within the package contains input and requests from the user. (Api, Controllers, Redirections, Resources)
+ *
+ * @author Julien Linget
+ */
+package fr.siamois.ui;
